@@ -31,8 +31,8 @@ android {
     testApplicationId "com.uber.autodispose.android.lifecycle.androidTest"
   }
   compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_7
-    targetCompatibility JavaVersion.VERSION_1_7
+    sourceCompatibility JavaVersion.VERSION_1_8
+    targetCompatibility JavaVersion.VERSION_1_8
   }
   lintOptions {
     lintConfig file('lint.xml')
