@@ -43,12 +43,7 @@ public abstract class MainThreadDisposable implements Disposable {
       if (AutoDisposeAndroidUtil.isMainThread()) {
         onDispose();
       } else {
-        AndroidSchedulers.mainThread().scheduleDirect(new Runnable() {
-          @Override
-          public void run() {
-            onDispose();
-          }
-        });
+        AndroidSchedulers.mainThread().scheduleDirect(this::onDispose);
       }
     }
   }
