@@ -19,7 +19,6 @@ package com.ubercab.autodispose.rxlifecycle;
 import com.trello.rxlifecycle2.LifecycleProvider;
 import com.trello.rxlifecycle2.OutsideLifecycleException;
 import com.uber.autodispose.ScopeProvider;
-import io.reactivex.CompletableSource;
 
 /**
  * Interop for RxLifecycle. This provides static factory methods to convert {@link
@@ -49,13 +48,9 @@ public final class RxLifecycleInterop {
    * @return a {@link ScopeProvider}
    */
   public static <E> ScopeProvider from(final LifecycleProvider<E> provider) {
-    return new ScopeProvider() {
-      @Override public CompletableSource requestScope() {
-        return provider.lifecycle()
-            .compose(provider.bindToLifecycle())
-            .ignoreElements();
-      }
-    };
+    return () -> provider.lifecycle()
+        .compose(provider.bindToLifecycle())
+        .ignoreElements();
   }
 
   /**
@@ -74,12 +69,8 @@ public final class RxLifecycleInterop {
    * @return a {@link ScopeProvider}
    */
   public static <E> ScopeProvider from(final LifecycleProvider<E> provider, final E event) {
-    return new ScopeProvider() {
-      @Override public CompletableSource requestScope() {
-        return provider.lifecycle()
-            .compose(provider.bindUntilEvent(event))
-            .ignoreElements();
-      }
-    };
+    return () -> provider.lifecycle()
+        .compose(provider.bindUntilEvent(event))
+        .ignoreElements();
   }
 }
